serialize/deserialize Block with RLP tested

+++ libs/common/bit_array.go
@@ -11,7 +11,7 @@ import (
 // BitArray is a thread-safe implementation of a bit array.
 type BitArray struct {
 	mtx   sync.Mutex
-	Bits  int      `json:"bits"`  // NOTE: persisted via reflect, must be exported
+	Bits  uint32   `json:"bits"`  // NOTE: persisted via reflect, must be exported
 	Elems []uint64 `json:"elems"` // NOTE: persisted via reflect, must be exported
 }
 
@@ -22,7 +22,7 @@ func NewBitArray(bits int) *BitArray {
 		return nil
 	}
 	return &BitArray{
-		Bits:  bits,
+		Bits:  uint32(bits),
 		Elems: make([]uint64, (bits+63)/64),
 	}
 }
@@ -32,7 +32,7 @@ func (bA *BitArray) Size() int {
 	if bA == nil {
 		return 0
 	}
-	return bA.Bits
+	return int(bA.Bits)
 }
 
 // GetIndex returns the bit at index i within the bit array.
@@ -47,7 +47,7 @@ func (bA *BitArray) GetIndex(i int) bool {
 }
 
 func (bA *BitArray) getIndex(i int) bool {
-	if i >= bA.Bits {
+	if i >= int(bA.Bits) {
 		return false
 	}
 	return bA.Elems[i/64]&(uint64(1)<<uint(i%64)) > 0
@@ -65,7 +65,7 @@ func (bA *BitArray) SetIndex(i int, v bool) bool {
 }
 
 func (bA *BitArray) setIndex(i int, v bool) bool {
-	if i >= bA.Bits {
+	if i >= int(bA.Bits) {
 		return false
 	}
 	if v {
@@ -99,7 +99,7 @@ func (bA *BitArray) copyBits(bits int) *BitArray {
 	c := make([]uint64, (bits+63)/64)
 	copy(c, bA.Elems)
 	return &BitArray{
-		Bits:  bits,
+		Bits:  uint32(bits),
 		Elems: c,
 	}
 }
@@ -123,7 +123,7 @@ func (bA *BitArray) Or(o *BitArray) *BitArray {
 		bA.mtx.Unlock()
 		o.mtx.Unlock()
 	}()
-	c := bA.copyBits(MaxInt(bA.Bits, o.Bits))
+	c := bA.copyBits(MaxInt(int(bA.Bits), int(o.Bits)))
 	for i := 0; i < len(c.Elems); i++ {
 		c.Elems[i] |= o.Elems[i]
 	}
@@ -147,7 +147,7 @@ func (bA *BitArray) And(o *BitArray) *BitArray {
 }
 
 func (bA *BitArray) and(o *BitArray) *BitArray {
-	c := bA.copyBits(MinInt(bA.Bits, o.Bits))
+	c := bA.copyBits(MinInt(int(bA.Bits), int(o.Bits)))
 	for i := 0; i < len(c.Elems); i++ {
 		c.Elems[i] &= o.Elems[i]
 	}
@@ -193,7 +193,7 @@ func (bA *BitArray) Sub(o *BitArray) *BitArray {
 		}
 		i := len(o.Elems) - 1
 		if i >= 0 {
-			for idx := i * 64; idx < o.Bits; idx++ {
+			for idx := i * 64; idx < int(o.Bits); idx++ {
 				c.setIndex(idx, c.getIndex(idx) && !o.getIndex(idx))
 			}
 		}
@@ -271,9 +271,9 @@ func (bA *BitArray) PickRandom() (int, bool) {
 			if elemBits == 0 {
 				elemBits = 64
 			}
-			randBitStart := RandIntn(elemBits)
-			for j := 0; j < elemBits; j++ {
-				bitIdx := ((j + randBitStart) % elemBits)
+			randBitStart := RandIntn(int(elemBits))
+			for j := 0; j < int(elemBits); j++ {
+				bitIdx := ((j + randBitStart) % int(elemBits))
 				if (bA.Elems[elemIdx] & (uint64(1) << uint(bitIdx))) > 0 {
 					return 64*elemIdx + bitIdx, true
 				}
@@ -307,7 +307,7 @@ func (bA *BitArray) StringIndented(indent string) string {
 func (bA *BitArray) stringIndented(indent string) string {
 	lines := []string{}
 	bits := ""
-	for i := 0; i < bA.Bits; i++ {
+	for i := 0; i < int(bA.Bits); i++ {
 		if bA.getIndex(i) {
 			bits += "x"
 		} else {
@@ -372,7 +372,7 @@ func (bA *BitArray) MarshalJSON() ([]byte, error) {
 	defer bA.mtx.Unlock()
 
 	bits := `"`
-	for i := 0; i < bA.Bits; i++ {
+	for i := 0; i < int(bA.Bits); i++ {
 		if bA.getIndex(i) {
 			bits += `x`
 		} else {
